Split current tests into multiple files
